==> 0-1. VISITJEJU_CRAWLING.ipynb <==
# 비짓제주 크롤링 코드

비짓제주 숙박/관광지/F&B/쇼핑 데이터 크롤링 코드


from bs4 import BeautifulSoup
from selenium import webdriver
import re
import warnings
import pandas as pd

warnings.filterwarnings(action='ignore')

base_url = 'https://www.visitjeju.net/kr/detail/list?menuId=DOM_000001707000000000&cate1cd=cate0000000004#p1&pageSize=12&sortListType=reviewcnt&viewType=thumb&isShowBtag&tag' #보여지는 형태를 꼭 네모 네개 짜리로 바꿀 것

driver = webdriver.Chrome(executable_path = 'D:\Python_417\chromedriver')
driver.get(url=base_url)
# implicit Waits
driver.implicitly_wait(time_to_wait = 5)

cleaning = re.compile('<p.*="">')

html = driver.page_source
soup = BeautifulSoup(html, 'html.parser')
places = soup.select('dt>a>p.s_tit') # 제목
places = [cleaning.sub('', str(name))[:-4] for name in places]
# plc_tag = soup.find_all('p', class_= 'item_tag prev')
# plc_tag = [(cleaning.sub('', str(tags))[:-4]).strip() for tags in plc_tag]
# plc_tag는 안에 들어가면 더 있기 때문에 들어가서 구하는 걸로 하자

tmp = driver.find_element_by_css_selector('#content > div > div.cont_wrap > div.recommend_area > ul > li:nth-child(1) > dl > dt > a')
# 여기의 li:nth-child  (1) 안의 숫자를 바꿔가면서 클릭하면 됨
tmp.click()

## (1) 비짓제주 숙박 데이터 크롤링

def click_get_info():
    twice = False
    try:
        name = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.sub_info_title > h3').text
        stars = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.score_area_l > p').text[
              9:-1]
        tags = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.tag_area').text.replace(
            '\n', ' ')
        address = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.basic_information > div:nth-child(2) > p.info_sub_cont').text
        telephone = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.basic_information > div:nth-child(3) > p.info_sub_cont').text
        like = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > ul > li:nth-child(1) > button > p.appraisal_cnt').text
        heart = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > ul > li:nth-child(2) > button > p.appraisal_cnt').text
        numreview = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > ul > li:nth-child(3) > p.appraisal_cnt').text
        view = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > ul > li:nth-child(6) > p.appraisal_cnt').text
        ### 상세정보
        detail_text = driver.find_element_by_css_selector(
            '#tab0 > div').text.replace('\n', ' ')
        ### 추천콘텐츠
        #### 관광지
        rec_tour = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        #### 음식점
        try:
            driver.find_element_by_css_selector(
                '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_restaurant > a').click()
            rec_restaurant = driver.find_element_by_css_selector(
                '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        except:
            driver.find_element_by_css_selector('#tab0 > h2').click()
            driver.find_element_by_css_selector(
                '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_restaurant > a').click()
            rec_restaurant = driver.find_element_by_css_selector(
                '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
            twice=True
        #### 숙박
        driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_lodgment > a').click()
        rec_lodgment = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        #### 빅데이터
        driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_bigdata > a').click()
        rec_bigdata = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        ### 이용안내
        detail_side_box = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(2) > div > dl').text.split('\n')
    except:
        name = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.sub_info_title > h3').text
        stars = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.score_area_l > p').text[9:-1]
        tags = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.tag_area').text.replace('\n', ' ')
        tags+=' '+driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.tag_area > p:nth-child(2)').text.replace('\n','')
        address = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.basic_information > div:nth-child(2) > p.info_sub_cont').text
        telephone = driver.find_element_by_css_selector(
             '#content > div.cont.detail_page.detail_style2 > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.basic_information > div:nth-child(3) > p.info_sub_cont').text
        ## 좋아요
        like = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > ul > li:nth-child(1) > button > p.appraisal_cnt').text
        ### 찜하기
        heart = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > ul > li:nth-child(2) > button > p.appraisal_cnt').text
        ### 리뷰
        numreview = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > ul > li:nth-child(3) > p.appraisal_cnt').text
        ### 조회
        view = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > ul > li:nth-child(6) > p.appraisal_cnt').text
        ### 상세정보
        detail_text = driver.find_element_by_css_selector('#tab0 > div').text.replace('\n', ' ')
        ### 추천콘텐츠
        #### 관광지
        rec_tour = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        #### 음식점
        try:
            driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_restaurant > a').click()
            rec_restaurant = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        except:
            driver.find_element_by_css_selector('#tab0 > h2').click()
            driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_restaurant > a').click()
            rec_restaurant = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
            twice=True
        #### 숙박
        driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_lodgment > a').click()
        rec_lodgment = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        #### 빅데이터
        driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_bigdata > a').click()
        rec_bigdata = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        ### 이용안내
        detail_side_box = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(2) > div > dl').text.split('\n')

    detail_side_box = { detail_side_box[i] : detail_side_box[i+1] for i in range(0,len(detail_side_box), 2) }

    return {'name' : name, 'stars' : stars, 'tags': tags, 'address' : address, 'telephone' : telephone, 'like' : like, 'heart' : heart, 'numreview' : numreview, 'view' : view, 'detail_text' : detail_text, 'rec_tour' : rec_tour, 'rec_restaurant' : rec_restaurant, 'rec_lodgment' :  rec_lodgment, 'rec_bigdata' : rec_bigdata, 'detail_side_box' : detail_side_box}, twice

driver.get(url=base_url)
driver.maximize_window()
driver.implicitly_wait(time_to_wait = 5)
cleaning = re.compile('<p.*="">')
final_df = pd.DataFrame(columns = ['name', 'stars', 'tags', 'address', 'telephone', 'like', 'heart', 'numreview', 'view', 'detail_text', 'rec_tour',
                                   'rec_restaurant', 'rec_lodgment', 'rec_bigdata', 'detail_side_box'])
page_number = int(driver.find_element_by_css_selector('#content > div > div.cont_wrap > div.recommend_area > div.theme_tit_area.clear > div > div > div.counter > p > strong').text)//12

def crawling_tour(driver, final_df):
    places = [element.text.split('\n')[1] for element in driver.find_elements_by_css_selector('#content > div > div.cont_wrap > div.recommend_area > ul > li')]
    for i in range(1,len(places)+1):
        place_click = driver.find_element_by_css_selector(f'#content > div > div.cont_wrap > div.recommend_area > ul > li:nth-child({i}) > dl > dt > a')
        place_click.click()
        driver.implicitly_wait(time_to_wait = 20)

        tmp_info_dict, twice = click_get_info()
        tmp_df = pd.DataFrame(columns = tmp_info_dict.keys(), data = [tmp_info_dict.values()])
        final_df = final_df.append(tmp_df)

        driver.execute_script("window.history.go(-1)")
        if twice==True:
            driver.execute_script("window.history.go(-1)")
    # driver.execute_script("window.history.go(-1)")
    return final_df

for i in range(1,page_number):
    final_df = crawling_tour(driver, final_df)
    driver.find_element_by_css_selector('#paging > a.spr_com.page-next').click()
    driver.implicitly_wait(time_to_wait=10)
    print(f'{i}번째 페이지 끝')
    final_df

final_df = final_df.drop_duplicates('name')

# store_df = pd.read_csv('VISITJEJU_숙소정보크롤링.csv')

store_df = store_df.append(final_df)
store_df = store_df.drop_duplicates('name')
store_df.reset_index(drop=True, inplace=True)

#store_df.to_csv('data/VISITJEJU_숙박정보크롤링_FINAL.csv', encoding='utf-8-sig', index=False)

### 크롤링 안된 경우 체크

#store_df = pd.read_csv('bigcon_data/VISITJEJU_숙박정보크롤링.csv')

driver.get(url = base_url)

main_url = base_url

for i in range(1,80):
    driver.get(url = main_url)
    places = [element.text.split('\n')[1] for element in driver.find_elements_by_css_selector('#content > div > div.cont_wrap > div.recommend_area > ul > li')]
    print(places)
    main_url = driver.current_url
    for idx, place in enumerate(places):
        if place not in store_df['name'].tolist():
            place_click = driver.find_element_by_css_selector(f'#content > div > div.cont_wrap > div.recommend_area > ul > li:nth-child({idx+1}) > dl > dt > a')
            place_click.click()
            driver.implicitly_wait(time_to_wait = 20)

            tmp_info_dict, twice = click_get_info()
            tmp_df = pd.DataFrame(columns = tmp_info_dict.keys(), data = [tmp_info_dict.values()])
            store_df = store_df.append(tmp_df)

            driver.execute_script("window.history.go(-1)")
            if twice==True:
                driver.execute_script("window.history.go(-1)")

            print(place)
        else:
            pass
    driver.get(url = main_url)
    driver.find_element_by_css_selector('#paging > a.spr_com.page-next').click()
    driver.implicitly_wait(time_to_wait=60)
    main_url = driver.current_url
    print(f'{i}번째 페이지 끝')

store_df['name'].value_counts()

store_df[store_df['name'] == '해뜨는집']

- 해뜨는집이 visitjeju 숙박에 두개 있는데 두개가 하나는 남원, 하나는 도두에 위치하고 있어서 다르게 인식됨 그런데 이름에는 해뜨는집으로 공통되게 나오는 문제가 있음
- 그래서 하나를 해뜨는집 (남원)으로, 다른 걸 해뜨는집 (도두)로 바꿔주려고 함

store_df.at[282, 'name'] = '해뜨는집 (남원)'

store_df.reset_index(drop=True, inplace=True)
store_df[store_df['name'] == '해뜨는집']

s = store_df.copy()

store_df.at[954, 'name' ] = '해뜨는집 (도두)'
store_df = store_df.drop(953, axis=0)
store_df = store_df.drop_duplicates('name').reset_index(drop=True)

store_df.to_csv('data/VISITJEJU_숙박크롤링_FINAL.csv', encoding='utf-8-sig', index=False)

## (2) 비짓제주 관광지 데이터 크롤링

url = 'https://www.visitjeju.net/kr/detail/list?menuId=DOM_000001707000000000&cate1cd=cate0000000004#p1&pageSize=12&sortListType=reviewcnt&viewType=thumb&isShowBtag&tag' #보여지는 형태를 꼭 네모 네개 짜리로 바꿀 것

driver = webdriver.Chrome(executable_path = 'D:\Python_417\chromedriver')
driver.get(url=url)

# implicit Waits
driver.implicitly_wait(time_to_wait = 5)

cleaning = re.compile('<p.*="">')

html = driver.page_source
soup = BeautifulSoup(html, 'html.parser')
places = soup.select('dt>a>p.s_tit') # 제목
places = [cleaning.sub('', str(name))[:-4] for name in places]
# plc_tag = soup.find_all('p', class_= 'item_tag prev')
# plc_tag = [(cleaning.sub('', str(tags))[:-4]).strip() for tags in plc_tag]
# plc_tag는 안에 들어가면 더 있기 때문에 들어가서 구하는 걸로 하자

tmp = driver.find_element_by_css_selector('#content > div > div.cont_wrap > div.recommend_area > ul > li:nth-child(1) > dl > dt > a')
# 여기의 li:nth-child(1) 안의 숫자를 바꿔가면서 클릭하면 됨
tmp.click()

def click_get_info():
    twice = False
    try:
        name = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.sub_info_title > h3').text
        stars = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.score_area_l > p').text[
              9:-1]
        tags = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.tag_area').text.replace(
            '\n', ' ')
        address = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.basic_information > div:nth-child(2) > p.info_sub_cont').text
        telephone = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.basic_information > div:nth-child(3) > p.info_sub_cont').text
        like = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > ul > li:nth-child(1) > button > p.appraisal_cnt').text
        heart = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > ul > li:nth-child(2) > button > p.appraisal_cnt').text
        numreview = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > ul > li:nth-child(3) > p.appraisal_cnt').text
        view = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > ul > li:nth-child(6) > p.appraisal_cnt').text
        ### 상세정보
        detail_text = driver.find_element_by_css_selector(
            '#tab0 > div').text.replace('\n', ' ')
        ### 추천콘텐츠
        #### 관광지
        rec_tour = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        #### 음식점
        try:
            driver.find_element_by_css_selector(
                '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_restaurant > a').click()
            rec_restaurant = driver.find_element_by_css_selector(
                '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        except:
            driver.find_element_by_css_selector('#tab0 > h2').click()
            driver.find_element_by_css_selector(
                '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_restaurant > a').click()
            rec_restaurant = driver.find_element_by_css_selector(
                '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
            twice=True
        #### 숙박
        driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_lodgment > a').click()
        rec_lodgment = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        #### 빅데이터
        driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_bigdata > a').click()
        rec_bigdata = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        ### 이용안내
        detail_side_box = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(2) > div > dl').text.split('\n')
    except:
        name = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.sub_info_title > h3').text
        stars = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.score_area_l > p').text[9:-1]
        tags = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.tag_area').text.replace('\n', ' ')
        tags+=' '+driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.tag_area > p:nth-child(2)').text.replace('\n','')
        address = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.basic_information > div:nth-child(2) > p.info_sub_cont').text
        telephone = driver.find_element_by_css_selector(
             '#content > div.cont.detail_page.detail_style2 > div.sub_visual_wrap > div.inner_wrap > div.sub_info_area > div.basic_information > div:nth-child(3) > p.info_sub_cont').text
        ## 좋아요
        like = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > ul > li:nth-child(1) > button > p.appraisal_cnt').text
        ### 찜하기
        heart = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > ul > li:nth-child(2) > button > p.appraisal_cnt').text
        ### 리뷰
        numreview = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > ul > li:nth-child(3) > p.appraisal_cnt').text
        ### 조회
        view = driver.find_element_by_css_selector(
            '#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > ul > li:nth-child(6) > p.appraisal_cnt').text
        ### 상세정보
        detail_text = driver.find_element_by_css_selector('#tab0 > div').text.replace('\n', ' ')
        ### 추천콘텐츠
        #### 관광지
        rec_tour = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        #### 음식점
        try:
            driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_restaurant > a').click()
            rec_restaurant = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        except:
            driver.find_element_by_css_selector('#tab0 > h2').click()
            driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_restaurant > a').click()
            rec_restaurant = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
            twice=True
        #### 숙박
        driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_lodgment > a').click()
        rec_lodgment = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        #### 빅데이터
        driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_category > ul > li.add2020_detail_bigdata > a').click()
        rec_bigdata = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(1) > div.add2020_detail_side_contents').text.split('\n')[0:-1:2]
        ### 이용안내
        detail_side_box = driver.find_element_by_css_selector('#content > div.cont.detail_page.detail_style2 > div.cont_wrap.sub_visual > div.add2020_detail > div.add2020_detail_right > div:nth-child(2) > div > dl').text.split('\n')

    detail_side_box = { detail_side_box[i] : detail_side_box[i+1] for i in range(0,len(detail_side_box), 2) }

    return {'name' : name, 'stars' : stars, 'tags': tags, 'address' : address, 'telephone' : telephone, 'like' : like, 'heart' : heart, 'numreview' : numreview, 'view' : view, 'detail_text' : detail_text, 'rec_tour' : rec_tour, 'rec_restaurant' : rec_restaurant, 'rec_lodgment' :  rec_lodgment, 'rec_bigdata' : rec_bigdata, 'detail_side_box' : detail_side_box}, twice

driver.get(url=url)
driver.maximize_window()
driver.implicitly_wait(time_to_wait = 5)
cleaning = re.compile('<p.*="">')
final_df = pd.DataFrame(columns = ['name', 'stars', 'tags', 'address', 'telephone', 'like', 'heart', 'numreview', 'view', 'detail_text', 'rec_tour',
                                   'rec_restaurant', 'rec_lodgment', 'rec_bigdata', 'detail_side_box'])
page_number = int(driver.find_element_by_css_selector('#content > div > div.cont_wrap > div.recommend_area > div.theme_tit_area.clear > div > div > div.counter > p > strong').text)//12

def crawling_tour(driver, final_df):
    places = [element.text.split('\n')[1] for element in driver.find_elements_by_css_selector('#content > div > div.cont_wrap > div.recommend_area > ul > li')]
    for i in range(1,len(places)+1):
        place_click = driver.find_element_by_css_selector(f'#content > div > div.cont_wrap > div.recommend_area > ul > li:nth-child({i}) > dl > dt > a')
        place_click.click()
        driver.implicitly_wait(time_to_wait = 20)

        tmp_info_dict, twice = click_get_info()
        tmp_df = pd.DataFrame(columns = tmp_info_dict.keys(), data = [tmp_info_dict.values()])
        final_df = final_df.append(tmp_df)

        driver.execute_script("window.history.go(-1)")
        if twice==True:
            driver.execute_script("window.history.go(-1)")

        print(f'{i}번째 관광지')
    # driver.execute_script("window.history.go(-1)")
    return final_df

for i in range(2,page_number):
    final_df = crawling_tour(driver, final_df)
    driver.find_element_by_css_selector('#paging > a.spr_com.page-next').click()
    driver.implicitly_wait(time_to_wait=10)
    final_df

final_df = final_df.drop_duplicates('name').reset_index(drop=True)

final_df.to_csv('data/VISITJEJU_관광지정보크롤링_0914.csv', encoding='utf-8-sig', index=False)

## (3) 비짓제주 F&B 데이터 크롤링
