==> 0-2. VEGANMAP_CRAWLING.ipynb <==
# 제주비건지도 크롤링

from bs4 import BeautifulSoup

from selenium import webdriver

import re

import warnings
warnings.filterwarnings(action='ignore')

import pandas as pd

base_url = "http://www.jejuvegan.com/vegan_map"

driver = webdriver.Chrome(executable_path = 'D:\Python_417\chromedriver')
driver.get(url=base_url)

html = driver.page_source
soup = BeautifulSoup(html, 'html.parser')
driver.quit()

raw = soup.select('div.col-md-9.wz_title')

final_df = pd.DataFrame(columns=['title', 'option', 'phone', 'menu', 'work', 'off', 'address'])
final_df['sns'], final_df['etc'] = 'empty', 'empty'
final_df

def option_function (option):
    option = option[1:-2]
    if option == '/img/vegan_icon_s.png':
        option = '비건 식당'
    elif option == '/img/vegan_option_icon_s.png':
        option = '비건 옵션'
    elif option == '/img/dessert_s_icon.png':
        option = '비건 디저트'
    else:
        option = '베이커리'
    return option

gettext_function = lambda x : x.get_text().strip()

def info_function(content):
    basic_idx = [2,7,10,13]
    phone, sns, etc = 'empty', 'empty', 'empty'
    if len(content)<=16:
        if '연락처: ' == content[1].get_text():
            phone, menu, work, off = [gettext_function(content[x]) for x in basic_idx]
            address = gettext_function(content[15])
        else:
            menu, work, off, address = [gettext_function(content[x]) for x in [2,5,7,10]]
    elif len(content)<=21:
        phone, menu, work, off = [gettext_function(content[x]) for x in basic_idx]
        address = gettext_function(content[15])
        sns = content[18]['href']
    else: # len(content) >=21:
        phone, menu, work, off = [gettext_function(content[x]) for x in basic_idx]
        address, etc =  [gettext_function(content[x]) for x in [15, 18]]
        sns = content[20]['href']
    return phone, menu, work, off, address, sns, etc

 def vegan_crawling(line):
    # 식당 이름
    title = line.select('div > h3 > a')[0].get_text()
    # 식당 옵션 (비건 식당, 비건 옵션, 비건 디저트, 비건 베이커리)
    option = str(line.select('div > img')[0])
    option = re.search('".*" ', option).group()
    option = option_function(option)
    content = line.p.contents
    try:
        phone, menu, work, off, address, sns, etc = info_function(content)
    except:
        phone, menu, work, off, address, sns, etc = ['error'] * 7
    return {'title' : title, 'option' : option, 'phone' : phone,
            'menu' : menu, 'work' : work, 'off' : off, 'address' : address,
            'etc' : etc, 'sns' : sns}

for idx, line in enumerate(raw):
    print(idx)
    final_df = final_df.append(vegan_crawling(line), ignore_index=True)

final_df['phone'].value_counts() # error 3개 정도는 손으로 바꿔줌

final_df.to_csv('././bigcon_data/제주비건식당크롤링.csv', encoding='utf-8-sig', index=False)
