==> 0-4. RENTCAR_CRAWLING.ipynb <==
# 렌터카 가격 정보 크롤링

- 가능한 많은 정보를 받으려고 사이트 두 곳에서 크롤링
1. 올댓여행 렌터카 페이지 크롤링
2. 엔젤렌터카 신한카드 페이지 크롤링

from bs4 import BeautifulSoup

from selenium import webdriver
from selenium.webdriver.support.ui import WebDriverWait
import time

import re
import pandas as pd

import warnings
warnings.filterwarnings(action='ignore')

## 1. 올댓여행 렌터카 페이지 크롤링

url = 'https://allthat.travelbonus.co.kr/Main.asp?BIGDIV_VAL=RentCar#tabRentcar01'

path= "C://Users//hjh05//Desktop//mefasol//code//chromedriver.exe"
driver = webdriver.Chrome(path)
driver.get(url=url)

# implicit Waits
driver.implicitly_wait(time_to_wait = 5)

html = driver.page_source
soup = BeautifulSoup(html, 'html.parser')

driver.get(url=url)
driver.maximize_window()
driver.implicitly_wait(time_to_wait = 5)
cleaning = re.compile('<p.*="">')
final_df = pd.DataFrame(columns = ['SM_CAT', 'FUEL_TYPE', 'PRICE'])
car_type = 5

try :
    nomore = driver.find_element_by_css_selector('#Chk')
    nomore.click()
except : 
    print("no pop-up")

SM_CAT = []
FUEL_TYPE = []
PRICE = []

for j in range(1,6) :
    info = [element.text.split('\n') for element in driver.find_elements_by_css_selector(f'#over_rentcar0{j} > div')][0]    
    
    for k in range(0,len(info)//5):
        SM_CAT.append(info[k*5])
        FUEL_TYPE.append(info[k*5+1]) 
        PRICE.append(re.sub(",","",info[k*5+4]).split("원")[0])
    
    if j != 5 : 
        type_click = driver.find_element_by_css_selector(f'#over_rentcar0{j} > ul > li:nth-child({1+j})')
        type_click.click()

final_df = pd.DataFrame()
final_df['SM_CAT'] = SM_CAT
final_df['FUEL_TYPE'] = FUEL_TYPE
final_df['PRICE'] = PRICE

final_df

driver.quit()

## 2. 엔젤렌터카 신한카드 페이지 크롤링

path= "C://Users//hjh05//Desktop//mefasol//code//chromedriver.exe"
driver = webdriver.Chrome(path)

url = 'https://jejuangelcar.com/tour.php?p=realcar-list'
driver.get(url=url)
driver.maximize_window()

source = driver.page_source
soup = BeautifulSoup(source, 'html.parser')
cleaning = re.compile('<p.*="">')

driver.implicitly_wait(time_to_wait = 10)

#날짜 바꾸기
driver.find_element_by_css_selector('#l-wrap2 > section > form > div > div.searchInline.searchCalendar.datePicker > img').click()
#10월 18일
driver.find_element_by_css_selector('#renewalDatePicker > div > div.ui-datepicker-group.ui-datepicker-group-middle > table > tbody > tr:nth-child(4) > td:nth-child(3) > a').click()
#10월 19일
driver.find_element_by_css_selector('#renewalDatePicker > div > div.ui-datepicker-group.ui-datepicker-group-middle > table > tbody > tr:nth-child(4) > td:nth-child(4) > a').click()

driver.find_element_by_css_selector('#serach_form_submit').click()
driver.find_element_by_css_selector('#l-wrap2 > section > form > div > div.searchInline.searchButton > button').click()


# implicit Waits
driver.implicitly_wait(time_to_wait = 5)

final_df2 = pd.DataFrame(columns = ['SM_CAT', 'FUEL_TYPE', 'PRICE'])
car_type = 5

def click_get_info():
    SM_CAT = driver.find_element_by_css_selector('#info_table > div > div.carItem.w65 > div > p.itemTitle').text
    info = driver.find_element_by_css_selector('#info_table > div > div.carItem.w34 > p.itemTag').text
    FUEL_TYPE = info.split()[0]
    price = driver.find_element_by_css_selector('#info_table > div > div.carPrice > div.priceInline.w48.text-right > p.price > span').text

    return {'SM_CAT' : SM_CAT, 'FUEL_TYPE' : FUEL_TYPE, 'PRICE': price}

def scroll():
    SCROLL_PAUSE_TIME = 3
    last_height = driver.execute_script("return document.body.scrollHeight") 

    for j in range(6): #5번 
        driver.execute_script("window.scrollTo(0, document.body.scrollHeight);") 
        time.sleep(SCROLL_PAUSE_TIME) 

scroll()

cars = len([element.text.split('\n') for element in driver.find_elements_by_css_selector('#content > section > ul')][0])//7

for i in range(2,cars+5): #여유를 두고
    try :
        place_click = driver.find_element_by_css_selector(f'#content > section > ul > li:nth-child({i}) > div.buttonList > button.rentButton.direct.bookingButton.rentButtonOpen')
        place_click.click()
        driver.implicitly_wait(time_to_wait = 5)
    
        tmp_info_dict = click_get_info()
        tmp_df = pd.DataFrame(columns = tmp_info_dict.keys(), data = [tmp_info_dict.values()])
        final_df2 = final_df.append(tmp_df)

        driver.back()

        scroll()
    except :
        print("no more info",i)
        break

final_df2

driver.quit()

finals = pd.concat([final_df,final_df2], ignore_index=True)
finals

finals.to_csv('ALLTHAT_제주렌터카.csv', encoding='utf-8-sig', index=False)
