==> 1-0. CO2_VISITJEJULOG_PREPROCESSING.ipynb <==
# 사전 전처리

숙박/관광지/F&B 별 데이터셋을 만들기 위해 지번주소 단위 온실가스 배출량 데이터(TB_ECO_BUILDING_GHG_GIS_JEJU_1920)와 비짓제주 조회수 데이터(JT_MT_ACCTO_TRRSRT_SCCNT_LIST)를 전처리하는 코드입니다.

1. 지번주소 단위 온실가스 배출 데이터((TB_ECO_BUILDING_GHG_GIS_JEJU_1920) 전처리
 - 월별 탄소배출량 변수 생성
 - 월별 일자로 나누어 일별 탄소배출량으로 계산
 - 지번주소별 탄소배출량 평균/최댓값 계산
   
  
 2. 비짓제주 조회수 데이터(JT_MT_ACCTO_TRRSRT_SCCNT_LIST) 전처리
 - 0으로 채워진 결측치 NA로 변경
 - 이름 전처리
 - 월별 조회 수가 다른 행으로 나눠져 있는 경우 정리


import pandas as pd
import requests
import logging
from tqdm import tqdm

## 1. 온실가스 배출량 데이터 전처리


### 1-(1) 월별 배출량 변수 생성

data = pd.read_csv('data/TB_ECO_BUILDING_GHG_GIS_JEJU_1920.csv', encoding='cp949')

data = data.groupby(['LTNO_ADDR', 'USE_YR', 'USE_MONTH']).sum().reset_index()
data['YEAR_MONTH'] = data['USE_YR'].apply(str) + '_' + data['USE_MONTH'].apply(str)

data = data[['LTNO_ADDR', 'GPS_LNGTD', 'GPS_LTTD', 'SUM_GRGS_DSAMT', 'YEAR_MONTH']]

data

data_unique = data[['LTNO_ADDR', 'GPS_LNGTD', 'GPS_LTTD']].drop_duplicates().reset_index(drop=True) ; data_unique

data_unique[['2019_3', '2019_4', '2019_5', '2019_9', '2019_10', '2019_11', '2020_3', '2020_4', '2020_5', '2020_9', '2020_10']]= 0.0

for value, yr_mt,add in zip(data['SUM_GRGS_DSAMT'],data['YEAR_MONTH'], data['LTNO_ADDR']):
    du_idx = data_unique[data_unique['LTNO_ADDR'] == add].index[0]
    data_unique.at[du_idx, yr_mt] = value

data_unique = data_unique[['LTNO_ADDR', 'GPS_LNGTD', 'GPS_LTTD', '2019_3', '2019_4', '2019_5','2019_9', '2019_10', '2019_11', '2020_3', '2020_4', '2020_5', '2020_9', '2020_10']]

### 1-(2) 일별 데이터로 계산
월별 일자 수로 각 월별 탄소배출량 나누기

data = data_unique.copy()

data['2019_3'] = data['2019_3']/31
data['2019_4'] = data['2019_4']/30
data['2019_5'] = data['2019_5']/31
data['2019_9'] = data['2019_9']/30
data['2019_10'] = data['2019_10']/31
data['2019_11'] = data['2019_11']/30
data['2020_3'] = data['2020_3']/31
data['2020_4'] = data['2020_4']/30
data['2020_5'] = data['2020_5']/31
data['2020_9'] = data['2020_9']/30
data['2020_10'] = data['2020_10']/31

### 1-(3) 지번주소별 탄소배출량 평균/최댓값 계산

data['co2_mean'] = [ data.iloc[idx][data.columns[-11:]].mean() for idx in range(len(data))]
data['co2_max'] = [ data.iloc[idx][data.columns[-11:]].max() for idx in range(len(data))]

data = data.drop(['2019_3', '2019_4', '2019_5', '2019_9', '2019_10', '2019_11', '2020_3', '2020_4', '2020_5', '2020_9', '2020_10'], axis=1)
data.to_csv('data/PRE_CO2_지번단위탄소배출량_meanmax.csv', encoding='utf-8-sig', index=False)

data

## 2. 비짓제주 조회수 데이터 전처리

import pandas as pd
import numpy as np
import re

def name_preprocessing(name):
    name = name.replace('(주)','').replace("·",".").lower()
    name = re.sub('[㈜!@#$%^&*{}|:;\'\"/><,~`+=_ ]','', name)
    name = re.sub('[\[.\]]','',name)
    name = name.split('(')[0]
    return name.strip()

def vj_grouping(x):
    return pd.Series({'AREA_NM' : x['AREA_NM'].unique(),
                      'ADDR' : x['ADDR'].unique(),
                      'BASE_YEAR' : x['BASE_YEAR'].sum(),
                      'MAR_VIEW_CO' : int(x['MAR_VIEWS_CO'].sum()),
                      'APR_VIEW_CO' : int(x['APR_VIEWS_CO'].sum()),
                      'MAY_VIEW_CO' : int(x['MAY_VIEWS_CO'].sum()),
                      'SEP_VIEW_CO' : int(x['SEP_VIEWS_CO'].sum()),
                      'OCT_VIEW_CO' : int(x['OCT_VIEWS_CO'].sum()),
                      'NOV_VIEW_CO' : int(x['NOV_VIEWS_CO'].sum())})

vj_view = pd.read_csv('data/JT_MT_ACCTO_TRRSRT_SCCNT_LIST.csv', encoding='cp949')

# 0.0으로 채워져 있는 행들을 다 NA로 바꿈
for hmm in vj_view.columns:
    vj_view[f'{hmm}'] = vj_view[f'{hmm}'].replace(0.0, np.nan)

vj_view

# 이름 preprocessing -> 새로운 열 만듦
vj_view['AREA_NM_PRE'] = vj_view['AREA_NM'].apply(name_preprocessing)

# 3월/4월/5월/9월/10월/11월이 각각 다른 행으로 잡혀 있는 거 수정
vj_view = vj_view.groupby(['AREA_NM_PRE', 'CL_NM']).apply(vj_grouping)
vj_view = vj_view.reset_index() ; vj_view

vj_view.to_csv('data/PRE_JT_MT_ACCTO_TRRSRT_SCCNT_LIST.csv', encoding='utf-8-sig', index=False)
