==> 6.PROTOTYPE_FUNCTIONAL.ipynb <==
import pandas as pd
import googlemaps
from haversine import haversine

google_api = "###################"

class UserInput():
    def __init__(self, time_stamp=1667178000):
        self.time_stamp = time_stamp  # Unix Timestamp (default is 2022년 10월 31일 10:00 GMT )
        dep_address = {'김해국제공항': '부산광역시 강서구 공항진입로', '김포국제공항': '서울특별시 강서구 하늘길 112',
                       '인천국제공항': '인천광역시 중구 공항로 272', '목포항': '전라남도 목포시 해안로 182 목포항연안여객터미널',
                       '제주공항': '제주특별자치도 제주시 공항로 2', '제주항': '제주특별자치도 제주시 임항로 111'}

        self.before_trip_start = input('여행을 시작하는 곳의 도로명 주소를 입력해주세요. "서울특별시 중구 을지로 100"과 같은 도로명 주소 형태로 입력해주세요.')
        self.before_trip_seoul = input(
            '어느 공항/선착장에서 출발하시나요? /n 김포국제공항, 인천국제공항, 김해국제공항, 목포항 중에 선택 해주세요. \n 정확한 공항/선착장의 이름을 입력해주세요.')
        self.before_trip_jeju = input('어느 공항/선착장으로 도착하시나요? /n 제주공항, 제주항 중에 선택해주세요.\n 정확한 공항/선착장의 이름을 입력해주세요.')
        self.before_trip_how = input('제주도로 가는 교통편을 선택해주세요. /n 비행기, 배')

        # 주소로 매핑해줍니다
        self.before_trip_seoul = dep_address[self.before_trip_seoul]
        self.before_trip_jeju = dep_address[self.before_trip_jeju]

        lodgement = pd.read_csv('co2_calc/탄소발자국계산기_숙소_FINAL.csv')
        self.before_trip_date = int(input("제주에 머무는 기간을 알려주세요. '3박 4일'의 형태로 입력해주세요.").split(' ')[1][:-1])

        print(lodgement['type'].unique().tolist())
        self.before_trip_lod_type = input('어느 유형의 숙소에 묵으실 예정인가요?')

        print(lodgement[lodgement['type'] == self.before_trip_lod_type]['AREA_NM_PRE'])
        self.before_trip_lod_name = input('어느 숙소에 묵으실 건가요? 정확한 이름을 입력해주세요.')
        try:
            self.lod_addr = lodgement[lodgement['AREA_NM_PRE'] == self.before_trip_lod_name]['지번주소'].values[0]
        except:
            self.before_trip_lod_name = input('숙소 이름이 정확하지 않습니다. 이름을 다시 입력해주세요.')
            self.lod_addr = lodgement[lodgement['AREA_NM_PRE'] == self.before_trip_lod_name]['ADDR'].values[0]

        self.lod_co2 = lodgement[lodgement['AREA_NM_PRE'] == self.before_trip_lod_name]['co2_max_객실'].values[0]

        self.after_trip_jeju = input('어느 공항/선착장에서 제주도를 떠나시나요? /n 제주공항, 제주항 \n 정확한 공항/항구의 이름을 입력해주세요. ')
        self.after_trip_seoul = input('어느 공항/선착장으로 도착하실 건가요? /n 김포국제공항, 인천국제공항, 김해국제공항, 목포항 \n 정확한 공항/선착장의 이름을 입력해주세요.')
        self.after_trip_how = input('제주도를 떠나는 교통편을 선택해주세요. /n 1. 비행기 2. 배')
        self.after_trip_end = input('여행이 끝나고 돌아가는 곳의 도로명 주소를 입력해주세요.')

        # 여기도 주소로 매핑해줍니다
        self.after_trip_jeju = dep_address[self.after_trip_jeju]
        self.after_trip_seoul = dep_address[self.after_trip_seoul]

    def day_trip(self):
        """
        매일 매일의 여정을 받아주는 함수입니다.
        3박 4일의 경우 4회의 입력을 받습니다.
        관광지/F&B를 선택하는 과정 없이 목적지와 목적지로 가는 교통편을 입력받습니다.
        최종 목적지로 숙소/공항/선착장을 입력하는 경우 일정을 받는 과정이 종료됩니다.
        :return:
        """
        visit_list, trans_list = [], []
        stop_flag = False
        while not stop_flag:
            where_you_go = input("목적지를 입력해주세요.")
            how_you_go = input(
                "목적지까지 가는 방법을 입력해주세요. \n 도보, 대중교통, 자전거, 택시, 렌트카 중에서 입력해주세요. /n 렌트카의 경우 '렌트카'와 함께 정확한 차종명을 입력해주세요. (예시: 렌트카, K5) *콤마주의")

            visit_list.append(where_you_go)
            trans_list.append(how_you_go)

            if (where_you_go == "숙소") | (where_you_go == '공항') | (where_you_go == '선착장'):
                break
        return visit_list, trans_list

    # ----------------------------------------------------------------------------------------------------

    def calc_dist_gmap(self, start, end, google_api):
        """
        구글맵API를 통해 start-end 사이의 거리와 소요 시간을 계산합니다.
        :param start: original location (도로명 주소)
        :param end: destination location (도로명 주소)
        :param google_api: api key
        :return:
        """
        gmaps = googlemaps.Client(key=google_api)
        try:
            result = gmaps.directions(start, end, mode='transit', departure_time=self.time_stamp, language='ko')
        except:
            return 'no_search', 'no_search', 'no_serach', 'no_search'
        try:
            result = result[0]['legs'][0]
        except:
            return 'dist_error', 'dist_error', 'dist_error', 'dist_error'
        howtogo = [step['travel_mode'] for step in result['steps']]
        return result['distance']['text'], result['duration']['text'], howtogo, result

    def get_lon_lat(self, place, google_api):
        gmaps = googlemaps.Client(key=google_api)
        result = gmaps.geocode((place), language='ko')
        return result[0]["geometry"]["location"]["lat"], result[0]["geometry"]["location"]["lng"]

    # -------------------------------------------------------------------------------------------------------

    def calc_trans_co2(self, date, visit_list, trans_list):
        """
        :param visit_list: ex) ['테디베어뮤지엄제주', '오설록티뮤지엄', '노형수퍼마켙', '엉또폭포', '숙소']
        :param trans_list: ex) ['대중교통', '택시', '렌트카, K5', '도보', '자전거'']
        :return:
        """
        data = pd.read_csv('co2_calc/탄소발자국계산기_데이터_FINAL.csv')
        trans_data = pd.read_csv('co2_calc/탄소발자국계산기_데이터_교통.csv')

        # 첫째날은 출발지를 공항/선착지로, 마지막날은 도착지를 공항/선착지로 한다
        lod_1, lod_2 = self.lod_addr, self.lod_addr
        if date == 'first':
            lod_1 = self.before_trip_jeju
        elif date == 'last':
            lod_2 = self.after_trip_jeju

        try:
            visit_addr = [lod_1] + [data[data['AREA_NM'] == visit]['address'].values[0] for visit in
                                    visit_list[:-1]] + [lod_2]
        except:
            return 'name_not_in_dataset'
        visit_list = [lod_1] + visit_list[:-1] + [lod_2]

        # 빈 리스트 생성해서 -> 여기다가 다 담아줄 것
        dist_list, times_list, howtogo_list, result_list = [], [], [], []

        # 거리, 시간, 어케 가는지, 결과 계산
        for idx, visit in enumerate(visit_addr[:-1]):
            dist, times, howtogo, result = self.calc_dist_gmap(visit_addr[idx], visit_addr[idx + 1], google_api)
            if dist == 'dist_error':
                print(f'{visit_list[idx]}와 {visit_list[idx + 1]}의 거리가 너무 멉니다. 다시 입력해주세요.')
                return 'dist_error'
            elif dist == 'no_search':
                print(f'{visit_list[idx]} 또는 {visit_list[idx]}가 검색이 되지 않습니다. 확인해주세요.')
                return 'no_search'
            dist_list.append(dist)
            times_list.append(times)
            howtogo_list.append(howtogo)
            result_list.append(result)

        # trans_list (어떻게갈건지)에 따라서 co2 계산
        co2_list_trans, co2 = [], 0.0

        for idx, trans in enumerate(trans_list):
            if (trans == '도보') | (trans == '자전거'):
                co2 = 0.0
            elif trans == '택시':
                co2 = (float(dist_list[idx][:-3]) * trans_data[trans_data['MD_CAT'] == '택시']['FACTOR'].values[0]) / 4
            elif trans[:3] == '렌트카':
                car_type = trans.split(' ')[1]
                trans_tmp = trans_data[trans_data['LG_CAT'] == '자동차'].reset_index(drop=True)
                trans_tmp = trans_data[trans_data['FUEL_TYPE'] == '휘발유'].reset_index(drop=True)
                for jdx, t in enumerate(trans_tmp['SM_CAT']):
                    if car_type in t:
                        co2 = (float(dist_list[idx][:-3]) * trans_tmp['FACTOR'][jdx]) // trans_tmp['N_PEOPLE'][jdx]
                        break
                co2 = 0.173463301 # 차종 제대로 입력 안 했을 경우 평균 값으로 넣어줌
            elif trans == '대중교통':
                co2 = 0.0
                for jdx, how in enumerate(howtogo_list[idx]):
                    if how == 'TRANSIT':
                        co2 += float(result_list[idx]['steps'][jdx]['distance']['text'][:-3]) *\
                        trans_data[trans_data['SM_CAT'] == '시내버스']['FACTOR'].values[0]

            co2_list_trans.append(co2)

        return co2_list_trans


    def calc_visit_co2(self, date, visit_list):
        data = pd.read_csv('co2_calc/탄소발자국계산기_데이터_FINAL.csv')

        co2_list_visit = []
        for visit in visit_list[:-1]:
            co2_list_visit.append(data[data['AREA_NM'] == visit]['co2'].values[0])
        if date != 'last':
            co2_list_visit.append(self.lod_co2)
        return co2_list_visit


    def calc_day_co2(self):
        """
        여행 중 발생하는 co2양을 매일 매일 계산 하는 함수입니다.
        :return:
        """
        date = [day for day in range(1, self.before_trip_date + 1)]
        date[0], date[-1] = 'first', 'last'
        self.all_co2_trans = []
        self.all_co2_visit = []
        self.all_co2_lod = float(self.lod_co2) * self.before_trip_date  # 숙박 일자로 곱해줌
        self.all_trans = []
        self.all_visit = []
        for day in date:
            print(f"-----------------{day}일 차 ")
            visit_list, trans_list = self.day_trip()
            self.all_trans.extend(trans_list)
            self.all_visit.extend(visit_list)
            tmp_trans = self.calc_trans_co2(day, visit_list, trans_list)
            if tmp_trans in ['car_error', 'dist_error', 'name_not_in_dataset', 'no_search']:
                if tmp_trans == 'car_error':
                    print("데이터에 없는 차종입니다. 다시 입력해주세요.")
                elif tmp_trans == 'dist_error':
                    print("거리가 너무 멀거나 가까운 조합이 있습니다. 다시 입력해주세요.")
                elif tmp_trans == 'name_not_in_dataset':
                    print("데이터에 없는 곳에 방문하셨습니다. 이름을 확인하시거나 다른 곳을 입력해주세요.")
                elif tmp_trans == 'no_search':
                    print("검색이 되지 않는 장소를 입력하셨습니다. 다시 한 번 확인해주세요.")
                return tmp_trans, tmp_trans, tmp_trans, tmp_trans, tmp_trans
            self.all_co2_trans.extend(self.calc_trans_co2(day, visit_list, trans_list))
            self.all_co2_visit.extend(self.calc_visit_co2(day, visit_list))
        return self.all_co2_trans, self.all_trans, self.all_co2_visit, self.all_visit, self.all_co2_lod


    def calc_all_co2(self):
        """
        여행 여정 중 발생하는 모든 co2양을 계산하는 함수입니다.
        :return:
        """
        trans_data = pd.read_csv('co2_calc/탄소발자국계산기_데이터_교통.csv')
        car_max = max(trans_data[trans_data['LG_CAT'] == '자동차']['FACTOR'])
        airplane = trans_data[trans_data['LG_CAT'] == '항공']['FACTOR'].values[0]
        ship = trans_data[trans_data['MD_CAT'] == '선박']['FACTOR'].values[0]

        # 시작지 -> 공항/선착장 까지의 거리 * 자동차
        start_coord = self.get_lon_lat(self.before_trip_start, google_api)
        before_coord = self.get_lon_lat(self.before_trip_seoul, google_api)
        before_dist = haversine(start_coord, before_coord, unit='km')
        before_co2 = int(before_dist * car_max)

        # (출발)공항/선착장 -> (도착)공항/선착장 까지의 거리 * 수단 탄소 배출량
        arrival_coord = self.get_lon_lat(self.before_trip_jeju, google_api)
        going_jeju_dist = haversine(before_coord, arrival_coord, unit='km')
        if self.before_trip_how == '비행기':
            going_jeju_co2 = going_jeju_dist * airplane
        else:
            going_jeju_co2 = going_jeju_dist * ship

        # 여행지에서 발생하는 탄소량 (교통, 방문지-관광지,FnB, 숙소*일자)

        date = [day for day in range(1, self.before_trip_date + 1)]
        date[0], date[-1] = 'first', 'last'
        self.all_co2_trans = []
        self.all_co2_visit = []
        self.all_co2_lod = float(self.lod_co2) * self.before_trip_date  # 숙박 일자로 곱해줌
        self.all_trans = []
        self.all_visit = []

        print(f'출발지인 {self.before_trip_start}에서 {self.before_trip_seoul}까지 자동차로 가는 동안 {before_co2}의 탄소가 배출됩니다.')
        print(
            f'{self.before_trip_seoul}에서 {self.before_trip_jeju}까지 {self.before_trip_how}로 가는 동안 {going_jeju_co2}의 탄소가 배출됩니다.')
        print(f'숙소 {self.before_trip_lod_name}에서 {self.before_trip_date - 1}일 동안 묵는 동안 {self.all_co2_lod}의 탄소가 배출됩니다.')

        for day in date:
            print(f"-----------------{day}일 차 입력 받는 중")
            visit_list, trans_list = self.day_trip()
            self.all_trans.extend(trans_list)
            self.all_visit.extend(visit_list)
            tmp_trans = self.calc_trans_co2(day, visit_list, trans_list)
            if tmp_trans in ['car_error', 'dist_error', 'name_not_in_dataset', 'no_search']:
                if tmp_trans == 'car_error':
                    print("데이터에 없는 차종입니다. 다시 입력해주세요.")
                elif tmp_trans == 'dist_error':
                    print("거리가 너무 먼 조합이 있습니다. 다시 입력해주세요.")
                elif tmp_trans == 'name_not_in_dataset':
                    print("데이터에 없는 곳에 방문하셨습니다. 이름을 확인하시거나 다른 곳을 입력해주세요.")
                elif tmp_trans == 'no_search':
                    print("검색이 되지 않는 장소를 입력하셨습니다. 다시 한 번 확인해주세요.")
                return tmp_trans, tmp_trans, tmp_trans, tmp_trans, tmp_trans
            self.all_co2_trans.extend(self.calc_trans_co2(day, visit_list, trans_list))
            self.all_co2_visit.extend(self.calc_visit_co2(day, visit_list))

        for check in self.all_co2_trans:
            if check in ['car_error', 'dist_error', 'name_not_in_dataset']:
                print('문제가 발생했습니다. 다시 한 번 확인하고 입력해주세요.')
                return check

        # (출발)공항/선착장 -> (도착)공항/선착장 까지의 거리 * 수단 탄소 배출량
        departure_coord = self.get_lon_lat(self.after_trip_jeju, google_api)
        after_coord = self.get_lon_lat(self.after_trip_seoul, google_api)
        leaving_jeju_dist = haversine(departure_coord, after_coord, unit='km')
        if self.after_trip_how == '비행기':
            leaving_jeju_co2 = leaving_jeju_dist * airplane
        else:
            leaving_jeju_co2 = leaving_jeju_dist * ship

        # (도착)공항/선착장 -> 도착지 까지의 거리 * 자동차
        end_coord = self.get_lon_lat(self.after_trip_end, google_api)
        after_dist = haversine(after_coord, end_coord, unit='km')
        after_co2 = after_dist * car_max

        for check in self.all_co2_visit:
            if check in ['car_error', 'dist_error', 'name_not_in_dataset']:
                print('문제가 발생했습니다. 다시 한 번 확인하고 입력해주세요.')
                return check

        print("------------------탄소영수증 발행 중---------------")
        self.all_co2_visit, self.all_visit = [0.0] + self.all_co2_visit, ['공항/선착장'] + self.all_visit
        for idx, (trans_co2, trans, visit_co2, visit) in enumerate(
                zip(self.all_co2_trans, self.all_trans, self.all_co2_visit[1:], self.all_visit[1:])):
            print(f'{self.all_visit[idx]}에서 {visit}(으)로 {trans}(으)로 이동하는 동안 {trans_co2}의 탄소가 배출됩니다.')
            print(f'{visit}에서는 {visit_co2}만큼의 탄소가 배출됩니다.')

        print('-------------------------------------------------')
        print(f"제주를 떠나며 {leaving_jeju_co2}의 탄소를 배출합니다.")
        print(f"{self.after_trip_end}로 향하며 {after_co2}의 탄소를 배출합니다.")
        print(f'총 {before_co2 + going_jeju_co2 + self.all_co2_lod + sum(self.all_co2_trans) + sum(self.all_co2_visit) + leaving_jeju_co2 + after_co2}의 탄소가 배출됩니다.')
        return before_co2, going_jeju_co2, self.all_co2_trans, self.all_trans, self.all_co2_visit, self.all_visit, self.all_co2_lod, leaving_jeju_co2, after_co2

## 입력예시:
- 서울특별시 중구 을지로 100
- 김포국제공항
- 제주공항
- 비행기
- 3박 4일
- 농어촌민박
- 하바다통나무집
- 제주공항
- 김포국제공항
- 비행기
- 서울특별시 중구 을지로 100


co2 = UserInput()

# 입력예시
- 숙소명: 제공된 탄소발자국계산기_숙소_FINAL.csv의 'AREA_NM_PRE'열의 값을 입력 해 주세요
- 관광지의 경우

co2.calc_all_co2()

#0
co2 = UserInput()
co2.calc_all_co2()

#1
co2 = UserInput()
co2.calc_all_co2()

#3
co2 = UserInput()
co2.calc_all_co2()

#5
co2 = UserInput()

co2.calc_all_co2()
